Tree and stem tables from Luquillo.
Source: R/data-luquillo-census.R

Tree and stem tables from Luquillo produced with rtbl 0.0.0.9000. These is how to interpret the names of the datasets:

The number following 'tree' or 'stem' indicates that the data corresponds to that specific census number.

Datasets without a number contain multiple censuses.

Datasets labeled with '1ha' contain all trees in only one hectare.

Datasets labeled with 'random' contain a total of 1,000 trees chosen at random from the entire plot.

Datasets labeled with 'tiny' are cherry-picked to produce a tiny set of data useful for most examples.

luquillo_tree6_random

luquillo_tree6_1ha

luquillo_stem_random

luquillo_stem_1ha

luquillo_stem6_1ha

luquillo_stem6_random

luquillo_stem_random_tiny

Format

See examples.

Details

For a general description see census_description.

See also

luquillo_vft_4quad, census_description, data_dictionary.

Examples

# Save space
short_str <- function(x) {
  str(x, give.attr = FALSE, list.len = 5)
}

short_str(luquillo_tree6_random)

#> 'data.frame':	1004 obs. of  19 variables:
#>  $ treeID   : int  104 119 180 602 631 647 1086 1144 1168 1380 ...
#>  $ stemID   : int  143 158 225 736 775 793 1339 1410 1438 1702 ...
#>  $ tag      : chr  "10009" "100104" "100171" "100649" ...
#>  $ StemTag  : chr  "10009" "100104" "100174" "100649" ...
#>  $ sp       : chr  "DACEXC" "MYRSPL" "CASARB" "GUAGUI" ...
#>   [list output truncated]

short_str(luquillo_tree6_1ha)

#> 'data.frame':	8016 obs. of  19 variables:
#>  $ treeID   : int  46 47 49 50 51 54 56 57 60 64 ...
#>  $ stemID   : int  NA 53 57 165123 NA NA NA NA NA NA ...
#>  $ tag      : chr  "100001" "100008" "100011" "100016" ...
#>  $ StemTag  : chr  NA "100008" "100011" "178258" ...
#>  $ sp       : chr  "PSYBRA" "PSYBRA" "PSYBRA" "PSYBRA" ...
#>   [list output truncated]

short_str(luquillo_stem_random)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	7944 obs. of  19 variables:
#>  $ treeID   : int  104 119 180 180 180 180 602 631 647 1086 ...
#>  $ stemID   : int  143 158 222 223 224 225 736 775 793 1339 ...
#>  $ tag      : chr  "10009" "100104" "100171" "100171" ...
#>  $ StemTag  : chr  "10009" "100104" "100095" "100096" ...
#>  $ sp       : chr  "DACEXC" "MYRSPL" "CASARB" "CASARB" ...
#>   [list output truncated]

short_str(luquillo_stem_1ha)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	72618 obs. of  19 variables:
#>  $ treeID   : int  46 47 47 47 47 47 47 47 47 47 ...
#>  $ stemID   : int  46 47 48 49 50 51 52 53 54 55 ...
#>  $ tag      : chr  "100001" "100008" "100008" "100008" ...
#>  $ StemTag  : chr  "100001" "100002" "100003" "100004" ...
#>  $ sp       : chr  "PSYBRA" "PSYBRA" "PSYBRA" "PSYBRA" ...
#>   [list output truncated]

short_str(luquillo_stem6_1ha)

#> 'data.frame':	12103 obs. of  19 variables:
#>  $ treeID   : int  46 47 47 47 47 47 47 47 47 47 ...
#>  $ stemID   : int  46 47 48 49 50 51 52 53 54 55 ...
#>  $ tag      : chr  "100001" "100008" "100008" "100008" ...
#>  $ StemTag  : chr  "100001" "100002" "100003" "100004" ...
#>  $ sp       : chr  "PSYBRA" "PSYBRA" "PSYBRA" "PSYBRA" ...
#>   [list output truncated]

short_str(luquillo_stem6_random)

#> 'data.frame':	1324 obs. of  19 variables:
#>  $ treeID   : int  104 119 180 180 180 180 602 631 647 1086 ...
#>  $ stemID   : int  143 158 222 223 224 225 736 775 793 1339 ...
#>  $ tag      : chr  "10009" "100104" "100171" "100171" ...
#>  $ StemTag  : chr  "10009" "100104" "100095" "100096" ...
#>  $ sp       : chr  "DACEXC" "MYRSPL" "CASARB" "CASARB" ...
#>   [list output truncated]

short_str(luquillo_stem_random_tiny)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	77 obs. of  19 variables:
#>  $ treeID   : int  10775 10863 23710 25331 63997 63997 63997 63997 63997 64070 ...
#>  $ stemID   : int  13496 13603 29640 31937 77760 77761 77762 77763 77764 77850 ...
#>  $ tag      : chr  "1128" "1129" "14434" "16164" ...
#>  $ StemTag  : chr  "1128" "1129" "14434" "16164" ...
#>  $ sp       : chr  "PREMON" "PREMON" "PREMON" "SLOBER" ...
#>   [list output truncated]