Data Dictionary of the ForestGEO database.
Source: R/data_dictionary.R

This dataset is based on the Tree Database Data Dictionary available at http://ctfs.si.edu/Public/DataDict/data_dict.php.

data_dictionary

Format

An object of class tbl_df (inherits from tbl, data.frame) with 242 rows and 3 columns.

Source

http://ctfs.si.edu/Public/DataDict/data_dict.php.

Examples

head(data_dictionary)

#> # A tibble: 6 x 3
#>   table  column           description                                          
#>   <chr>  <chr>            <chr>                                                
#> 1 Census CensusID         Primary key, an integer  automatically generated to ~
#> 2 Census PlotID           Foreign Key to Site table.                           
#> 3 Census PlotCensusNumber "Integer census number for an individual plot, 1=fir~
#> 4 Census StartDate        Date on which the first measurement of the census wa~
#> 5 Census EndDate          Date on which the last measurement of the census was~
#> 6 Census Description      Notes pertinent to the census or general description~

# Column names vary in subtle ways (e.g. SubSpeciesID vs. SubspeciesID)
pattern <- "^subspeciesid$"
matching <- grepl(pattern, data_dictionary$column, ignore.case = TRUE)
data_dictionary[matching, ]

#> # A tibble: 5 x 3
#>   table            column       description                                    
#>   <chr>            <chr>        <chr>                                          
#> 1 SpeciesInventory SubSpeciesID Foreign Key to  SubSpecies table.              
#> 2 Specimen         SubSpeciesID Foreign Key to SubSpecies table.               
#> 3 SubSpecies       SubSpeciesID Primary key, an integer  automatically generat~
#> 4 Tree             SubSpeciesID Foreign Key to  SubSpecies table, indicating t~
#> 5 ViewTaxonomy     SubspeciesID Foreign Key to  SubSpecies table, indicating t~