Complete ViewTaxonomy (taxa) table from Luquillo.

Compared to the raw data in text format, this dataset replaced NULL values with missing values, and its columns have a specific type.

luquillo_taxa

Format

An object of class tbl_df (inherits from tbl, data.frame) with 163 rows and 21 columns.

Acknowledgements

Thanks to Jess Zimmerman for sharing data from Luquillo.

See also

Examples

str(luquillo_taxa, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	163 obs. of  21 variables:
#>  $ ViewID        : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ SpeciesID     : int  56 57 58 59 60 61 62 63 64 65 ...
#>  $ SubspeciesID  : chr  NA NA NA NA ...
#>  $ Family        : chr  "Fabaceae" "Euphorbiaceae" "Euphorbiaceae" "Fabaceae-papilionoideae" ...
#>  $ Mnemonic      : chr  "AESAME" "ALCFLO" "ALCLAT" "ANDINE" ...
#>  $ Genus         : chr  "Aeschynomene" "Alchorneopsis" "Alchornea" "Andira" ...
#>  $ SpeciesName   : chr  "americana" "floribunda" "latifolia" "inermis" ...
#>  $ Rank          : chr  NA NA NA NA ...
#>  $ Subspecies    : chr  NA NA NA NA ...
#>  $ Authority     : chr  "(Poir.) Rudd" "(Benth.) Müll. Arg." "Sw." "(W. Wright) Kunth ex DC." ...
#>  $ IDLevel       : chr  "species" "species" "species" "species" ...
#>  $ subspMnemonic : chr  NA NA NA NA ...
#>  $ subspAuthority: chr  NA NA NA NA ...
#>  $ FieldFamily   : chr  "Fabaceae" "Euphorbiaceae" "Euphorbiaceae" "Fabaceae" ...
#>  $ Lifeform      : chr  NA NA NA NA ...
#>  $ Description   : chr  NA NA NA NA ...
#>  $ wsg           : num  0.678 0.427 0.42 0.642 0.636 ...
#>  $ wsglevel      : chr  "family" "species" "species" "species" ...
#>  $ ListOfOldNames: chr  NA NA NA NA ...
#>  $ Specimens     : chr  NA NA NA NA ...
#>  $ Reference     : chr  NA NA NA NA ...