Subset of a ViewFullTable (vft) from Luquillo.
Source: R/data-luquillo-view.R

Subset of a ViewFullTable (vft) from Luquillo. Included are all trees in four quadrats (coordinates: PX between (>=) 100 and 140; and PY between 400 and 440) of cencuses four through 6. For bigger datasets, including censuses one through six, see links under section See also.

luquillo_vft_4quad

Format

An object of class tbl_df (inherits from tbl, data.frame) with 3302 rows and 32 columns.

Details

Compared to the raw data in text format, this dataset replaced NULL values with missing values, and its columns have a specific type.

Acknowledgements

Thanks to Jess Zimmerman for sharing data from Luquillo.

See also

vft_description.

data_dictionary.

http://bit.ly/fgeo-data-luquillo-vft-random.

http://bit.ly/fgeo-data-luquillo-vft-1ha.

https://forestgeo.si.edu/sites/north-america/luquillo.

Examples

str(luquillo_vft_4quad, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	3302 obs. of  32 variables:
#>  $ DBHID           : int  384550 384553 466764 384554 466765 611446 384555 466766 611447 384556 ...
#>  $ PlotName        : chr  "luquillo" "luquillo" "luquillo" "luquillo" ...
#>  $ PlotID          : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Family          : chr  "Rubiaceae" "Urticaceae" "Urticaceae" "Urticaceae" ...
#>  $ Genus           : chr  "Psychotria" "Cecropia" "Cecropia" "Cecropia" ...
#>  $ SpeciesName     : chr  "berteroana" "schreberiana" "schreberiana" "schreberiana" ...
#>  $ Mnemonic        : chr  "PSYBER" "CECSCH" "CECSCH" "CECSCH" ...
#>  $ Subspecies      : chr  NA NA NA NA ...
#>  $ SpeciesID       : int  184 74 74 74 74 74 196 196 196 70 ...
#>  $ SubspeciesID    : chr  NA NA NA NA ...
#>  $ QuadratName     : chr  "721" "721" "721" "721" ...
#>  $ QuadratID       : int  327 327 327 327 327 327 327 327 327 327 ...
#>  $ PX              : num  127 130 130 127 127 ...
#>  $ PY              : num  404 405 405 406 406 ...
#>  $ QX              : num  6.98 9.51 9.51 6.95 6.95 6.95 9.88 9.88 9.88 8.56 ...
#>  $ QY              : num  4.11 4.53 4.53 5.74 5.74 5.74 8.94 8.94 8.94 9.08 ...
#>  $ TreeID          : int  692 698 698 706 706 706 711 711 711 714 ...
#>  $ Tag             : chr  "100768" "100774" "100774" "100781" ...
#>  $ StemID          : int  851 858 858 866 866 866 871 871 871 874 ...
#>  $ StemNumber      : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ StemTag         : int  100768 100774 100774 100781 100781 100781 100786 100786 100786 100789 ...
#>  $ PrimaryStem     : chr  NA NA NA NA ...
#>  $ CensusID        : int  4 4 5 4 5 6 4 5 6 4 ...
#>  $ PlotCensusNumber: int  4 4 5 4 5 6 4 5 6 4 ...
#>  $ DBH             : num  NA 185 NA 341 370 375 167 172 187 141 ...
#>  $ HOM             : num  NA 1.3 NA 1.28 1.3 1.95 1.3 1.3 1.43 1.3 ...
#>  $ ExactDate       : Date, format: "2006-11-21" "2006-11-21" ...
#>  $ Date            : int  17126 17126 19023 17125 19023 20663 17125 19023 20663 17125 ...
#>  $ ListOfTSM       : chr  "MAIN;DEADT" "MAIN;A" "MAIN;DEADT;R;ST" "MAIN;A" ...
#>  $ HighHOM         : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ LargeStem       : chr  NA NA NA NA ...
#>  $ Status          : chr  "dead" "alive" "dead" "alive" ...

show_some_values_of_each_column <- function(x) head(unique(x))
lapply(luquillo_vft_4quad, show_some_values_of_each_column)

#> $DBHID
#> [1] 384550 384553 466764 384554 466765 611446
#> 
#> $PlotName
#> [1] "luquillo"
#> 
#> $PlotID
#> [1] 1
#> 
#> $Family
#> [1] "Rubiaceae"  "Urticaceae" "Araliaceae" "Salicaceae" "Lauraceae" 
#> [6] "Meliaceae" 
#> 
#> $Genus
#> [1] "Psychotria" "Cecropia"   "Schefflera" "Casearia"   "Ocotea"    
#> [6] "Guarea"    
#> 
#> $SpeciesName
#> [1] "berteroana"   "schreberiana" "morototoni"   "arborea"      "leucoxylon"  
#> [6] "guidonia"    
#> 
#> $Mnemonic
#> [1] "PSYBER" "CECSCH" "SCHMOR" "CASARB" "OCOLEU" "GUAGUI"
#> 
#> $Subspecies
#> [1] NA
#> 
#> $SpeciesID
#> [1] 184  74 196  70 161 117
#> 
#> $SubspeciesID
#> [1] NA
#> 
#> $QuadratName
#> [1] "721" "621" "722" "622"
#> 
#> $QuadratID
#> [1] 327 326 343 342
#> 
#> $PX
#> [1] 126.98 129.51 126.95 129.88 128.56 126.50
#> 
#> $PY
#> [1] 404.11 404.53 405.74 408.94 409.08 411.04
#> 
#> $QX
#> [1] 6.98 9.51 6.95 9.88 8.56 6.50
#> 
#> $QY
#> [1]  4.11  4.53  5.74  8.94  9.08 11.04
#> 
#> $TreeID
#> [1] 692 698 706 711 714 725
#> 
#> $Tag
#> [1] "100768" "100774" "100781" "100786" "100789" "100799"
#> 
#> $StemID
#> [1] 851 858 866 871 874 885
#> 
#> $StemNumber
#> [1] 0
#> 
#> $StemTag
#> [1] 100768 100774 100781 100786 100789 100799
#> 
#> $PrimaryStem
#> [1] NA
#> 
#> $CensusID
#> [1] 4 5 6
#> 
#> $PlotCensusNumber
#> [1] 4 5 6
#> 
#> $DBH
#> [1]  NA 185 341 370 375 167
#> 
#> $HOM
#> [1]   NA 1.30 1.28 1.95 1.43 1.33
#> 
#> $ExactDate
#> [1] "2006-11-21" "2012-01-31" "2006-11-20" "2016-07-28" "2016-07-29"
#> [6] "2006-11-22"
#> 
#> $Date
#> [1] 17126 19023 17125 20663 20664 17127
#> 
#> $ListOfTSM
#> [1] "MAIN;DEADT"      "MAIN;A"          "MAIN;DEADT;R;ST" "MAIN;A;T"       
#> [5] "MAIN;A;C"        "MAIN;DEADT;B"   
#> 
#> $HighHOM
#> [1] 1
#> 
#> $LargeStem
#> [1] NA
#> 
#> $Status
#> [1] "dead"         "alive"        "broken below" "stem dead"   
#>