General description of a ViewFullTable (vft).
Source: R/data-view.R

Reproduced from http://bit.ly/vft-definitions:

ViewFullTable table. Merges all the relevant variables from the corresponding tables in the database back into a flat file. Each record represents one measurement and/or attribute of one stem of one tree of one plot from one census. Because buttresses can grow, for trees with buttresses, a stem may be measured at more than one height. So some stems may have more than one measurement in a census, but at different hom.

vft_description

Format

An object of class tbl_df (inherits from tbl, data.frame) with 32 rows and 2 columns.

See also

More information on columns definition: http://bit.ly/fgeo-tables-column-definition; http://bit.ly/fgeo-data-dictionary.

Examples

vft_description

#> # A tibble: 32 x 2
#>    column       definition                                                     
#>    <chr>        <chr>                                                          
#>  1 DBHID        Foreign Key to DBH table.                                      
#>  2 PlotName     Descriptive name of the site also referred to as the plot.     
#>  3 PlotID       Foreign Key to Site table.                                     
#>  4 Family       Taxonomic family name (from the Angiosperm Phylogeny Group - A~
#>  5 Genus        Genus of the plant, according to the APG system.               
#>  6 SpeciesName  Species part of Latin name, may be a morphospecies name.       
#>  7 Mnemonic     Code used in the field for designating the species, usually 6 ~
#>  8 Subspecies   Subspecies portion of the Latin name, may be a subspecies or v~
#>  9 SpeciesID    Foreign Key to Species table.                                  
#> 10 SubspeciesID Foreign Key to SubSpecies table, indicating the subspecies if ~
#> # ... with 22 more rows