General description of a ViewFullTable (vft).

Reproduced from http://bit.ly/vft-definitions:

ViewFullTable table. Merges all the relevant variables from the corresponding tables in the database back into a flat file. Each record represents one measurement and/or attribute of one stem of one tree of one plot from one census. Because buttresses can grow, for trees with buttresses, a stem may be measured at more than one height. So some stems may have more than one measurement in a census, but at different hom.

See also